+++ lib/log_flow/view/log_flow.dart
@@ -1,7 +1,7 @@
 import 'package:continual_care_alpha/app/bloc/app_bloc.dart';
+import 'package:continual_care_alpha/app/view/loading_page.dart';
 import 'package:continual_care_alpha/log_flow/log_flow.dart';
-import 'package:continual_care_alpha/log_flow/pages/badls_page.dart';
-import 'package:continual_care_alpha/log_flow/pages/iadls_page.dart';
+
 import 'package:flow_builder/flow_builder.dart';
 import 'package:flutter/material.dart';
 
@@ -16,6 +16,8 @@ List<Page> onGenerateLocationPages(LogState state, List<Page> pages) {
     if (state.status == LogStatus.tasksCompleted) IadlsPage.page(),
     if (state.status == LogStatus.iadlsCompleted) BadlsPage.page(),
     if (state.status == LogStatus.badlsCompleted) CommentsPage.page(),
+    if (state.status == LogStatus.loading) LoadingPage.page(),
+    if (state.status == LogStatus.success) LoadingPage.page(),
   ];
 }
