+++ lib/log_flow/widgets/widgets.dart
@@ -1,6 +1,7 @@
 export 'adl_check.dart';
 export 'date_ios_format.dart';
 export 'check_tile.dart';
+export 'comments.dart';
 export 'comment_tile.dart';
 export 'continue_button.dart';
 export 'main_title.dart';
